+++ packages/react/src/components/icon/icon.react.tsx
