+++ packages/react/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './button/button.react';
+export * from './icon/icon.react';
+export * from './input/input.react';
+export * from './textarea/textarea.react';
