+++ packages/react/src/components/search/search.react.tsx
@@ -1,8 +1,8 @@
 import { c, classy } from '@onfido/castor';
 import { Icon } from '@onfido/castor-react';
 import React from 'react';
+import { withRef } from '../../utils';
 import { Input, InputProps } from '../input/input.react';
-import { withRef } from '../utils';
 
 /**
  * @note `Search` uses an `Icon` that requires `Icons` (SVG sprite) to be
@@ -11,7 +11,10 @@ import { withRef } from '../utils';
  * https://github.com/onfido/castor-icons#use-with-plain-code
  */
 export const Search = withRef(
-  ({ className, style, ...restProps }: SearchProps, ref: SearchProps['ref']) => (
+  (
+    { className, style, ...restProps }: SearchProps,
+    ref: SearchProps['ref']
+  ) => (
     <div className={classy(c('search'), className)} style={style}>
       <Input {...restProps} ref={ref} type="search" />
       <Icon name="search" />
