+++ packages/react/src/index.ts
@@ -1,4 +1,1 @@
-export * from './button/button.react';
-export * from './icon/icon.react';
-export * from './input/input.react';
-export * from './textarea/textarea.react';
+export * from './components';
